fix : 삭제 API 오류 수정

+++ src/main/java/com/example/demo/module/review/service/ReviewServiceImpl.java
@@ -2,6 +2,8 @@ package com.example.demo.module.review.service;
 
 import com.example.demo.module.common.result.PageResult;
 import com.example.demo.module.hashtag.repository.HashtagRepository;
+import com.example.demo.module.image.entity.ReviewImage;
+import com.example.demo.module.image.service.ImageService;
 import com.example.demo.module.medicine.repository.MedicineRepository;
 import com.example.demo.module.point.entity.PointHistory;
 import com.example.demo.module.point.repository.PointHistoryRepository;
@@ -22,6 +24,8 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.io.IOException;
+import java.util.List;
 import java.util.NoSuchElementException;
 
 import static com.example.demo.module.point.entity.ReserveUse.RESERVE;
@@ -48,7 +52,7 @@ public class ReviewServiceImpl implements ReviewService {
         //TODO : 이미지 저장 로직 추가 필요
         User user = userRepository.findById(userId).orElseThrow(() -> new NoSuchElementException("해당하는 유저가 없습니다."));
 
-        if(reviewRepository.existsByMedicineIdAndUserUserId(reviewPayload.getMedicineId(), user.getUserId())){
+        if (reviewRepository.existsByMedicineIdAndUserUserId(reviewPayload.getMedicineId(), user.getUserId())) {
             throw new IllegalArgumentException("이미 후기를 작성한 영양제 입니다.");
         }
 
@@ -88,8 +92,11 @@ public class ReviewServiceImpl implements ReviewService {
     //이미지 수정은 따로 뺼 예정
     @Override
     @Transactional
-    public Long editReview(Long reviewId, ReviewEditPayload reviewEditPayload) {
-
+    public Long editReview(Long userId, Long reviewId, ReviewEditPayload reviewEditPayload) {
+        if (!reviewRepository.findById(reviewId).orElseThrow(() -> new NoSuchElementException("후기를 찾을 수 없습니다."))
+                .getUser().getUserId().equals(userId)) {
+            throw new IllegalArgumentException("후기 작성자만 수정할 수 있습니다.");
+        }
         //없어진 해쉬태그 삭제
         reviewHashtagRepository.findAllByReviewId(reviewId).forEach(ht -> {
             if (!reviewEditPayload.getTagList().contains(ht)) {
@@ -112,8 +119,13 @@ public class ReviewServiceImpl implements ReviewService {
         return review.update(reviewEditPayload);
     }
 
+    @Transactional
     @Override
-    public Long deleteByReviewId(Long reviewId) {
+    public Long deleteByReviewId(Long userId, Long reviewId) {
+        if (!reviewRepository.findById(reviewId).orElseThrow(() -> new NoSuchElementException("후기를 찾을 수 없습니다."))
+                .getUser().getUserId().equals(userId)) {
+            throw new IllegalArgumentException("후기 작성자만 삭제할 수 있습니다.");
+        }
         reviewRepository.deleteById(reviewId);
         return reviewId;
     }
